+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Included CLAUDE.md
 - Phase 0-2 complete
 - Phase 3 complete
+- Phase 4 complete